requirements.txt
datasets
psycopg2-binary
huggingface-hu